requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

pytest==8.3.4
httpx==0.27.2

google-auth==2.47.0
